history-setup.sh: Just nuke it
Y'know on second thought, we don't need that script at all.  The history
files themselves show the history -c and history -r commands, so there's
no need for this even as a method to teach people about them.  Plus,
we're going to be dispensing with the hsitory files as history files
anyway in the near future.