bash_history-rasppleii.sh: rc.local expansion
Perform the same whitespace expansion on the sed to modify /etc/rc.local
as has been performed in the history files.  The embedded code in the
sed command is now easier to read at the cost of legibility of the
history script itself.

This means that "running" this script no longer has 1:1 parity with
Ivan's history files, even if there is no functional difference.