bash_history-rasppleii.sh: Improve rc.local sed
Replace the && conditional in the rc.local sed with an if block.  Also,
comment the block so that the user has some idea why their config.txt is
being altered.  This is still not a great solution--better would be an
option to NOOBS not to alter the video mode at all.