bash_history-vm.sh: Don't modify /etc/sudoers
The /etc/sudoers file in Debian is a conffile.  That means if you change
it, your changes should be preserved across upgrades.  But also, it
means that any change to it will cause dpkg to stop and ask the user to
fix the conflict.

That should be avoided since the user didn't provide the change to the
conffile, we did.  And sudo (and a lot of other software nowadays) will
read files in /etc/sudoers.d/ allowing other packages or the user to
provide drop-in fragments that append or override the default
configuration.  That should be preferred, and we now do it that way.