bash_history-*: Document the history "commands"
This will be the last commit where the "commands" from Ivan's bash
history files are maintained in any way.  Before doing away with them,
it seemed only appropriate to document what they did exactly.

Also took a moment to translate the sed command into English.
Originally I wrote a long explanation of how this worked as a comment,
but the explanation was bigger than the file it was going into, so that
got removed.  I'll post the bigger explanation on Facebook.