motd.txt: Add unifide motd file
The changes between the Raspbian and VirtualBox Debian releases of
Raspple II come down to primarily branding.  The VirtualBox version has
simply been called A2SERVER virtual machine or similar, while Raspple II
is the port of that to Raspbian and the Raspberry Pi.

Once you unify the language a bit, the only real difference between them
is whether or not you tell the user about raspi-config, which only
exists on Raspbian.  That can be removed on Debian systems.