*.sh: Change tzdata to America/Los_Angeles
The default timezone has to be somewhere.  So much of what happens in
this country is based on Eastern time because New York and Washington DC
and all.  Admittedly what isn't is based on Pacific, but I happen to
live in Pacific, so I'm using it.  :P  America/Los_Angeles is correct
for Pacific timezone folks like me.

If you live somewhere else, you'll probably want to configure this on a
new installation anyway.